pyproject.toml
[project]
name = "cs-480"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dash>=4.0.0",
    "dotenv>=0.9.9",
    "einops>=0.8.2",
    "fastparquet>=2025.12.0",
    "google-genai>=1.64.0",
    "hdbscan>=0.8.41",
    "pdfminer>=20191125",
    "plotly>=6.5.2",
    "sentence-transformers>=5.2.2",
    "umap-learn>=0.5.11",
    "unstructured[pdf]>=0.18.32",
]
